[xxxx] Fixed misleading test
