+++ app/views/find/application/index.html.erb
@@ -1,9 +1,11 @@
+<% content_for :page_title, title_with_error_prefix(t("find.homepage.title"), @courses_by_location_or_training_provider_form.errors.present?) %>
+
 <div class="govuk-grid-row">
   <div class="govuk-grid-column-two-thirds-from-desktop">
     <%= form_with model: @courses_by_location_or_training_provider_form, url: find_age_groups_path, method: :post do |f| %>
       <%= f.govuk_error_summary %>
 
-      <%= f.govuk_radio_buttons_fieldset :find_courses, legend: { text: I18n.t("find.homepage.title"), size: "l" } do %>
+      <%= f.govuk_radio_buttons_fieldset :find_courses, legend: { text: t("find.homepage.title"), size: "l", tag: "h1" } do %>
         <%= f.govuk_radio_button :find_courses, "by_city_town_postcode", label: { text: t("find.homepage.radio_options.by_city_town_postcode.radio") }, link_errors: true do %>
           <%= f.govuk_text_field :find_courses, label: { text: t("find.homepage.radio_options.by_city_town_postcode.text_field") } %>
         <% end %>
