+++ spec/support/page_objects/find/courses_by_location_or_training_provider.rb
@@ -5,6 +5,8 @@ module PageObjects
     class CoursesByLocationOrTrainingProvider < PageObjects::Base
       set_url "/find"
 
+      element :heading, "h1"
+
       element :by_city_town_or_postcode_radio, "#find-courses-by-location-or-training-provider-form-find-courses-by-city-town-postcode-field"
       element :across_england, "#find-courses-by-location-or-training-provider-form-find-courses-across-england-field"
       element :by_school_uni_or_provider, "#find-courses-by-location-or-training-provider-form-find-courses-by-school-uni-or-provider-field"
